chore: extract logic in hook
